refactor: Move Task.js to component
