+++ src/js/components/Task.js
@@ -1,0 +1,21 @@
+import { TASK_DELAY } from "../utils/constans.js";
+
+class Task {
+	constructor(className) {
+		this.className = className;
+		this.element = document.querySelector(`.${className}`);
+	}
+
+	removeTask() {
+		this.element.innerHTML = "";
+	}
+
+	appendTask(calleeName) {
+		this.element.innerHTML = `<div class="task">${calleeName}</div>`;
+		setTimeout(() => {
+			this.removeTask();
+		}, TASK_DELAY);
+	}
+}
+
+export default Task;
